refactor(cli): rename commands
And use task `name` instead of `id` in user output
