+++ cmd/task/cmd.go
@@ -3,11 +3,9 @@ package task
 import (
 	"numerous.com/cli/cmd/args"
 	"numerous.com/cli/cmd/group"
+	"numerous.com/cli/cmd/task/instance"
 	"numerous.com/cli/cmd/task/instances"
 	"numerous.com/cli/cmd/task/list"
-	"numerous.com/cli/cmd/task/logs"
-	"numerous.com/cli/cmd/task/start"
-	"numerous.com/cli/cmd/task/stop"
 
 	"github.com/spf13/cobra"
 )
@@ -21,8 +19,6 @@ var Cmd = &cobra.Command{
 
 func init() {
 	Cmd.AddCommand(list.Cmd)
-	Cmd.AddCommand(start.Cmd)
+	Cmd.AddCommand(instance.Cmd)
 	Cmd.AddCommand(instances.Cmd)
-	Cmd.AddCommand(logs.Cmd)
-	Cmd.AddCommand(stop.Cmd)
 }
