+++ cmd/task/instance/create/mock.go
@@ -1,4 +1,4 @@
-package start
+package create
 
 import (
 	"context"
