+++ cmd/task/instance/create/start_test.go
@@ -1,4 +1,4 @@
-package start
+package create
 
 import (
 	"context"
