+++ cmd/task/instance/logs/mock.go
