+++ cmd/task/instance/stop/stop_test.go
